RF: Move constants into consts.py
Provide a place to define constants in, avoiding circular imports.

Closes #318
