+++ onyo/lib/consts.py
@@ -1,0 +1,4 @@
+# TODO: pseudo-key generation mapping
+NEW_PSEUDO_KEYS = ['path']  # TODO: name temporary b/c RF (old idea of pseudo-keys doesn't match)
+# The following keys have functional meaning for onyo and must not be part of user-defined asset content
+RESERVED_KEYS = ['directory', 'is_asset_directory', 'template']
